api_yamdb/requirements.txt
asgiref==3.3.2
attrs==22.2.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==2.0.12
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-csvimport==2.17
django-filter==22.1
django-import-export==3.0.2
django-model-utils==4.3.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt
et-xmlfile==1.1.0
idna==3.4
iniconfig==2.0.0
MarkupPy==1.14
mccabe==0.7.0
odfpy==1.4.1
openpyxl==3.1.0
packaging==23.0
pluggy==0.13.1
py==1.11.0
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==1.7.1
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2022.7.1
PyYAML==6.0
requests==2.26.0
sqlparse==0.4.3
tablib==3.3.0
toml==0.10.2
urllib3==1.26.14
xlrd==2.0.1
xlwt==1.3.0
gunicorn==20.0.4
psycopg2-binary==2.8.6